Add files via upload
Since the h5 trained model was too large to upload, I have instead just uploaded the github code. If run on a good system with a GPU and decent specs, the model shouldn't take more than 2 hours to train. The classifier model I created is highly accurate, and the code is such that the model is extremely scalable.